requirements.txt
email-validator==1.3.1
fastapi==0.89.1
PyJWT==2.6.0
python-decouple==3.7
uvicorn==0.20.0
pydantic[dotenv]~=1.10.7
motor~=3.1.2
passlib~=1.7.4
bcrypt~=4.0.1
pymongo[srv]~=4.3.3
